Rework impls.
