+++ crates/cache/src/engine.rs
@@ -1,6 +1,7 @@
 use crate::helpers::{is_writable, LOG_TARGET};
 use crate::items::{CacheItem, ProjectsState, RunTargetState, WorkspaceState};
 use crate::runfiles::CacheRunfile;
+use moon_archive::tar;
 use moon_constants::CONFIG_DIRNAME;
 use moon_error::MoonError;
 use moon_logger::{color, debug, trace};
@@ -14,13 +15,13 @@ pub struct CacheEngine {
     /// Contains cached items pertaining to runs and processes.
     pub dir: PathBuf,
 
-    /// The `.moon/cache/hashes` directory. Stores hash contents.
+    /// The `.moon/cache/hashes` directory. Stores hash manifests.
     pub hashes_dir: PathBuf,
 
     /// The `.moon/cache/runs` directory. Stores run states and runfiles.
     pub runs_dir: PathBuf,
 
-    /// The `.moon/cache/out` directory. Stores task output.
+    /// The `.moon/cache/out` directory. Stores task outputs as hashed archives.
     pub outputs_dir: PathBuf,
 }
 
@@ -82,6 +83,58 @@ impl CacheEngine {
         .await
     }
 
+    pub async fn create_hash_archive(
+        &self,
+        hash: &str,
+        project_root: &Path,
+        outputs: &[String],
+    ) -> Result<(), MoonError> {
+        if is_writable() {
+            // Old implementation would copy files to a hashed folder,
+            // so if we encounter that folder, let's just remove it!
+            let old_hash_folder = self.outputs_dir.join(hash);
+
+            if old_hash_folder.exists() && old_hash_folder.is_dir() {
+                fs::remove_dir_all(old_hash_folder).await?;
+            }
+
+            // New implementation uses tar archives! Very cool.
+            tar(
+                project_root,
+                outputs,
+                self.get_hash_archive_path(hash),
+                None,
+            )?;
+
+            trace!(
+                target: LOG_TARGET,
+                "Archiving outputs for {}",
+                color::symbol(hash)
+            );
+        }
+
+        Ok(())
+    }
+
+    pub async fn create_hash_manifest<T>(&self, hash: &str, hasher: &T) -> Result<(), MoonError>
+    where
+        T: ?Sized + Serialize,
+    {
+        if is_writable() {
+            let path = self.get_hash_manifest_path(hash);
+
+            trace!(
+                target: LOG_TARGET,
+                "Writing hash manifest {}",
+                color::path(&path)
+            );
+
+            fs::write_json(&path, &hasher, true).await?;
+        }
+
+        Ok(())
+    }
+
     pub async fn create_runfile<T: DeserializeOwned + Serialize>(
         &self,
         project_id: &str,
@@ -92,26 +145,21 @@ impl CacheEngine {
 
     pub async fn delete_hash(&self, hash: &str) -> Result<(), MoonError> {
         if is_writable() {
-            let path = self.get_hash_path(hash);
+            trace!(target: LOG_TARGET, "Deleting hash {}", color::symbol(hash));
 
-            trace!(target: "moon:cache:hash", "Deleting hash {}", color::path(&path));
-
-            // Remove the hash file itself
-            fs::remove_file(&path).await?;
-
-            // And the output with the hash
-            fs::remove_dir_all(&self.get_hash_output_dir(hash)).await?;
+            fs::remove_file(self.get_hash_manifest_path(hash)).await?;
+            fs::remove_file(self.get_hash_archive_path(hash)).await?;
         }
 
         Ok(())
     }
 
-    pub fn get_hash_path(&self, hash: &str) -> PathBuf {
-        self.hashes_dir.join(format!("{}.json", hash))
+    pub fn get_hash_archive_path(&self, hash: &str) -> PathBuf {
+        self.outputs_dir.join(format!("{}.tar.gz", hash))
     }
 
-    pub fn get_hash_output_dir(&self, hash: &str) -> PathBuf {
-        self.outputs_dir.join(hash)
+    pub fn get_hash_manifest_path(&self, hash: &str) -> PathBuf {
+        self.hashes_dir.join(format!("{}.json", hash))
     }
 
     pub fn get_project_dir(&self, project_id: &str) -> PathBuf {
@@ -122,632 +170,5 @@ impl CacheEngine {
         let path: PathBuf = [&target_id.replace(':', "/")].iter().collect();
 
         self.runs_dir.join(path)
-    }
-
-    pub async fn copy_output_to_out(
-        &self,
-        hash: &str,
-        source_root: &Path,
-        output_path: &Path,
-    ) -> Result<(), MoonError> {
-        if is_writable() {
-            let dest_root = self.get_hash_output_dir(hash);
-
-            fs::create_dir_all(&dest_root).await?;
-
-            trace!(
-                target: LOG_TARGET,
-                "Copying output {} to {}",
-                color::path(output_path),
-                color::path(&dest_root)
-            );
-
-            if output_path.is_file() {
-                fs::copy_file(
-                    output_path,
-                    dest_root.join(output_path.strip_prefix(source_root).unwrap()),
-                )
-                .await?;
-            } else {
-                fs::copy_dir_all(source_root, output_path, &dest_root).await?;
-            }
-        }
-
-        Ok(())
-    }
-
-    pub async fn save_hash<T>(&self, hash: &str, hasher: &T) -> Result<(), MoonError>
-    where
-        T: ?Sized + Serialize,
-    {
-        if is_writable() {
-            let path = self.get_hash_path(hash);
-
-            trace!(target: "moon:cache:hash", "Writing hash {}", color::path(&path));
-
-            fs::write_json(&path, &hasher, true).await?;
-        }
-
-        Ok(())
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::helpers::{run_with_env, to_millis};
-    use assert_fs::prelude::*;
-    use serial_test::serial;
-    use std::fs;
-
-    mod create {
-        use super::*;
-
-        #[tokio::test]
-        #[serial]
-        async fn creates_dirs() {
-            let dir = assert_fs::TempDir::new().unwrap();
-
-            CacheEngine::create(dir.path()).await.unwrap();
-
-            assert!(dir.path().join(".moon/cache").exists());
-            assert!(dir.path().join(".moon/cache/hashes").exists());
-            assert!(dir.path().join(".moon/cache/runs").exists());
-            assert!(dir.path().join(".moon/cache/out").exists());
-
-            dir.close().unwrap();
-        }
-    }
-
-    mod delete_hash {
-        use super::*;
-
-        #[tokio::test]
-        #[serial]
-        async fn deletes_files() {
-            let dir = assert_fs::TempDir::new().unwrap();
-            let cache = CacheEngine::create(dir.path()).await.unwrap();
-
-            dir.child(".moon/cache/hashes/abc123.json")
-                .write_str("{}")
-                .unwrap();
-
-            dir.child(".moon/cache/out/abc123/file.js")
-                .write_str("")
-                .unwrap();
-
-            let hash_file = cache.hashes_dir.join("abc123.json");
-            let out_file = cache.outputs_dir.join("abc123/file.js");
-
-            assert!(hash_file.exists());
-            assert!(out_file.exists());
-
-            cache.delete_hash("abc123").await.unwrap();
-
-            assert!(!hash_file.exists());
-            assert!(!out_file.exists());
-
-            dir.close().unwrap();
-        }
-
-        #[tokio::test]
-        #[serial]
-        async fn doesnt_delete_if_cache_off() {
-            let dir = assert_fs::TempDir::new().unwrap();
-            let cache = CacheEngine::create(dir.path()).await.unwrap();
-
-            dir.child(".moon/cache/hashes/abc123.json")
-                .write_str("{}")
-                .unwrap();
-
-            dir.child(".moon/cache/out/abc123/file.js")
-                .write_str("")
-                .unwrap();
-
-            let hash_file = cache.hashes_dir.join("abc123.json");
-            let out_file = cache.outputs_dir.join("abc123/file.js");
-
-            assert!(hash_file.exists());
-            assert!(out_file.exists());
-
-            run_with_env("off", || cache.delete_hash("abc123"))
-                .await
-                .unwrap();
-
-            assert!(hash_file.exists());
-            assert!(out_file.exists());
-
-            dir.close().unwrap();
-        }
-
-        #[tokio::test]
-        #[serial]
-        async fn doesnt_delete_if_cache_readonly() {
-            let dir = assert_fs::TempDir::new().unwrap();
-            let cache = CacheEngine::create(dir.path()).await.unwrap();
-
-            dir.child(".moon/cache/hashes/abc123.json")
-                .write_str("{}")
-                .unwrap();
-
-            dir.child(".moon/cache/out/abc123/file.js")
-                .write_str("")
-                .unwrap();
-
-            let hash_file = cache.hashes_dir.join("abc123.json");
-            let out_file = cache.outputs_dir.join("abc123/file.js");
-
-            assert!(hash_file.exists());
-            assert!(out_file.exists());
-
-            run_with_env("read", || cache.delete_hash("abc123"))
-                .await
-                .unwrap();
-
-            assert!(hash_file.exists());
-            assert!(out_file.exists());
-
-            dir.close().unwrap();
-        }
-    }
-
-    mod create_runfile {
-        use super::*;
-
-        #[tokio::test]
-        #[serial]
-        async fn creates_runfile_on_call() {
-            let dir = assert_fs::TempDir::new().unwrap();
-            let cache = CacheEngine::create(dir.path()).await.unwrap();
-            let runfile = cache
-                .create_runfile("123", &"content".to_owned())
-                .await
-                .unwrap();
-
-            assert!(runfile.path.exists());
-
-            assert_eq!(
-                fs::read_to_string(dir.path().join(".moon/cache/runs/123/runfile.json")).unwrap(),
-                "\"content\""
-            );
-
-            dir.close().unwrap();
-        }
-    }
-
-    mod cache_run_target_state {
-        use super::*;
-
-        #[tokio::test]
-        #[serial]
-        async fn creates_parent_dir_on_call() {
-            let dir = assert_fs::TempDir::new().unwrap();
-            let cache = CacheEngine::create(dir.path()).await.unwrap();
-            let item = cache.cache_run_target_state("foo:bar").await.unwrap();
-
-            assert!(!item.path.exists());
-            assert!(item.path.parent().unwrap().exists());
-
-            dir.close().unwrap();
-        }
-
-        #[tokio::test]
-        #[serial]
-        async fn loads_cache_if_it_exists() {
-            let dir = assert_fs::TempDir::new().unwrap();
-
-            dir.child(".moon/cache/runs/foo/bar/lastRunState.json")
-                .write_str(r#"{"exitCode":123,"hash":"","lastRunTime":0,"stderr":"","stdout":"","target":"foo:bar"}"#)
-                .unwrap();
-
-            let cache = CacheEngine::create(dir.path()).await.unwrap();
-            let item = cache.cache_run_target_state("foo:bar").await.unwrap();
-
-            assert_eq!(
-                item.item,
-                RunTargetState {
-                    exit_code: 123,
-                    target: String::from("foo:bar"),
-                    ..RunTargetState::default()
-                }
-            );
-
-            dir.close().unwrap();
-        }
-
-        #[tokio::test]
-        #[serial]
-        async fn loads_cache_if_it_exists_and_cache_is_readonly() {
-            let dir = assert_fs::TempDir::new().unwrap();
-
-            dir.child(".moon/cache/runs/foo/bar/lastRunState.json")
-                .write_str(r#"{"exitCode":123,"hash":"","lastRunTime":0,"stderr":"","stdout":"","target":"foo:bar"}"#)
-                .unwrap();
-
-            let cache = CacheEngine::create(dir.path()).await.unwrap();
-            let item = run_with_env("read", || cache.cache_run_target_state("foo:bar"))
-                .await
-                .unwrap();
-
-            assert_eq!(
-                item.item,
-                RunTargetState {
-                    exit_code: 123,
-                    target: String::from("foo:bar"),
-                    ..RunTargetState::default()
-                }
-            );
-
-            dir.close().unwrap();
-        }
-
-        #[tokio::test]
-        #[serial]
-        async fn doesnt_load_if_it_exists_but_cache_is_off() {
-            let dir = assert_fs::TempDir::new().unwrap();
-
-            dir.child(".moon/cache/runs/foo/bar/lastRunState.json")
-                .write_str(r#"{"exitCode":123,"hash":"","lastRunTime":0,"stderr":"","stdout":"","target":"foo:bar"}"#)
-                .unwrap();
-
-            let cache = CacheEngine::create(dir.path()).await.unwrap();
-            let item = run_with_env("off", || cache.cache_run_target_state("foo:bar"))
-                .await
-                .unwrap();
-
-            assert_eq!(
-                item.item,
-                RunTargetState {
-                    target: String::from("foo:bar"),
-                    ..RunTargetState::default()
-                }
-            );
-
-            dir.close().unwrap();
-        }
-
-        #[tokio::test]
-        #[serial]
-        async fn saves_to_cache() {
-            let dir = assert_fs::TempDir::new().unwrap();
-            let cache = CacheEngine::create(dir.path()).await.unwrap();
-            let mut item = cache.cache_run_target_state("foo:bar").await.unwrap();
-
-            item.item.exit_code = 123;
-
-            run_with_env("", || item.save()).await.unwrap();
-
-            assert_eq!(
-                fs::read_to_string(item.path).unwrap(),
-                r#"{"exitCode":123,"hash":"","lastRunTime":0,"stderr":"","stdout":"","target":"foo:bar"}"#
-            );
-
-            dir.close().unwrap();
-        }
-
-        #[tokio::test]
-        #[serial]
-        async fn doesnt_save_if_cache_off() {
-            let dir = assert_fs::TempDir::new().unwrap();
-            let cache = CacheEngine::create(dir.path()).await.unwrap();
-            let mut item = cache.cache_run_target_state("foo:bar").await.unwrap();
-
-            item.item.exit_code = 123;
-
-            run_with_env("off", || item.save()).await.unwrap();
-
-            assert!(!item.path.exists());
-
-            dir.close().unwrap();
-        }
-
-        #[tokio::test]
-        #[serial]
-        async fn doesnt_save_if_cache_readonly() {
-            let dir = assert_fs::TempDir::new().unwrap();
-            let cache = CacheEngine::create(dir.path()).await.unwrap();
-            let mut item = cache.cache_run_target_state("foo:bar").await.unwrap();
-
-            item.item.exit_code = 123;
-
-            run_with_env("read", || item.save()).await.unwrap();
-
-            assert!(!item.path.exists());
-
-            dir.close().unwrap();
-        }
-    }
-
-    mod cache_workspace_state {
-        use super::*;
-
-        #[tokio::test]
-        #[serial]
-        async fn creates_parent_dir_on_call() {
-            let dir = assert_fs::TempDir::new().unwrap();
-            let cache = CacheEngine::create(dir.path()).await.unwrap();
-            let item = cache.cache_workspace_state().await.unwrap();
-
-            assert!(!item.path.exists());
-            assert!(item.path.parent().unwrap().exists());
-
-            dir.close().unwrap();
-        }
-
-        #[tokio::test]
-        #[serial]
-        async fn loads_cache_if_it_exists() {
-            let dir = assert_fs::TempDir::new().unwrap();
-
-            dir.child(".moon/cache/workspaceState.json")
-                .write_str(r#"{"lastNodeInstallTime":123}"#)
-                .unwrap();
-
-            let cache = CacheEngine::create(dir.path()).await.unwrap();
-            let item = cache.cache_workspace_state().await.unwrap();
-
-            assert_eq!(
-                item.item,
-                WorkspaceState {
-                    last_node_install_time: 123,
-                    last_version_check_time: 0,
-                }
-            );
-
-            dir.close().unwrap();
-        }
-
-        #[tokio::test]
-        #[serial]
-        async fn loads_cache_if_it_exists_and_cache_is_readonly() {
-            let dir = assert_fs::TempDir::new().unwrap();
-
-            dir.child(".moon/cache/workspaceState.json")
-                .write_str(r#"{"lastNodeInstallTime":123}"#)
-                .unwrap();
-
-            let cache = CacheEngine::create(dir.path()).await.unwrap();
-            let item = run_with_env("read", || cache.cache_workspace_state())
-                .await
-                .unwrap();
-
-            assert_eq!(
-                item.item,
-                WorkspaceState {
-                    last_node_install_time: 123,
-                    last_version_check_time: 0,
-                }
-            );
-
-            dir.close().unwrap();
-        }
-
-        #[tokio::test]
-        #[serial]
-        async fn doesnt_load_if_it_exists_but_cache_is_off() {
-            let dir = assert_fs::TempDir::new().unwrap();
-
-            dir.child(".moon/cache/workspaceState.json")
-                .write_str(r#"{"lastNodeInstallTime":123}"#)
-                .unwrap();
-
-            let cache = CacheEngine::create(dir.path()).await.unwrap();
-            let item = run_with_env("off", || cache.cache_workspace_state())
-                .await
-                .unwrap();
-
-            assert_eq!(item.item, WorkspaceState::default());
-
-            dir.close().unwrap();
-        }
-
-        #[tokio::test]
-        #[serial]
-        async fn saves_to_cache() {
-            let dir = assert_fs::TempDir::new().unwrap();
-            let cache = CacheEngine::create(dir.path()).await.unwrap();
-            let mut item = cache.cache_workspace_state().await.unwrap();
-
-            item.item.last_node_install_time = 123;
-
-            run_with_env("", || item.save()).await.unwrap();
-
-            assert_eq!(
-                fs::read_to_string(item.path).unwrap(),
-                r#"{"lastNodeInstallTime":123,"lastVersionCheckTime":0}"#
-            );
-
-            dir.close().unwrap();
-        }
-    }
-
-    mod cache_projects_state {
-        use super::*;
-        use filetime::{set_file_mtime, FileTime};
-        use moon_utils::string_vec;
-        use std::collections::HashMap;
-        use std::time::SystemTime;
-
-        #[tokio::test]
-        #[serial]
-        async fn creates_parent_dir_on_call() {
-            let dir = assert_fs::TempDir::new().unwrap();
-            let cache = CacheEngine::create(dir.path()).await.unwrap();
-            let item = cache.cache_projects_state().await.unwrap();
-
-            assert!(!item.path.exists());
-            assert!(item.path.parent().unwrap().exists());
-
-            dir.close().unwrap();
-        }
-
-        #[tokio::test]
-        #[serial]
-        async fn loads_cache_if_it_exists() {
-            let dir = assert_fs::TempDir::new().unwrap();
-
-            dir.child(".moon/cache/projectsState.json")
-                .write_str(r#"{"globs":["**/*"],"projects":{"foo":"bar"}}"#)
-                .unwrap();
-
-            let cache = CacheEngine::create(dir.path()).await.unwrap();
-            let item = cache.cache_projects_state().await.unwrap();
-
-            assert_eq!(
-                item.item,
-                ProjectsState {
-                    globs: string_vec!["**/*"],
-                    projects: HashMap::from([("foo".to_owned(), "bar".to_owned())]),
-                }
-            );
-
-            dir.close().unwrap();
-        }
-
-        #[tokio::test]
-        #[serial]
-        async fn loads_cache_if_it_exists_and_cache_is_readonly() {
-            let dir = assert_fs::TempDir::new().unwrap();
-
-            dir.child(".moon/cache/projectsState.json")
-                .write_str(r#"{"globs":["**/*"],"projects":{"foo":"bar"}}"#)
-                .unwrap();
-
-            let cache = CacheEngine::create(dir.path()).await.unwrap();
-            let item = run_with_env("read", || cache.cache_projects_state())
-                .await
-                .unwrap();
-
-            assert_eq!(
-                item.item,
-                ProjectsState {
-                    globs: string_vec!["**/*"],
-                    projects: HashMap::from([("foo".to_owned(), "bar".to_owned())]),
-                }
-            );
-
-            dir.close().unwrap();
-        }
-
-        #[tokio::test]
-        #[serial]
-        async fn doesnt_load_if_it_exists_but_cache_is_off() {
-            let dir = assert_fs::TempDir::new().unwrap();
-
-            dir.child(".moon/cache/projectsState.json")
-                .write_str(r#"{"globs":[],"projects":{"foo":"bar"}}"#)
-                .unwrap();
-
-            let cache = CacheEngine::create(dir.path()).await.unwrap();
-            let item = run_with_env("off", || cache.cache_projects_state())
-                .await
-                .unwrap();
-
-            assert_eq!(item.item, ProjectsState::default());
-
-            dir.close().unwrap();
-        }
-
-        #[tokio::test]
-        #[serial]
-        async fn doesnt_load_if_it_exists_but_cache_is_stale() {
-            let dir = assert_fs::TempDir::new().unwrap();
-
-            dir.child(".moon/cache/projectsState.json")
-                .write_str(r#"{"globs":[],"projects":{"foo":"bar"}}"#)
-                .unwrap();
-
-            let now = to_millis(SystemTime::now()) - 100000;
-
-            set_file_mtime(
-                dir.path().join(".moon/cache/projectsState.json"),
-                FileTime::from_unix_time((now / 1000) as i64, 0),
-            )
-            .unwrap();
-
-            let cache = CacheEngine::create(dir.path()).await.unwrap();
-            let item = cache.cache_projects_state().await.unwrap();
-
-            assert_eq!(item.item, ProjectsState::default());
-
-            dir.close().unwrap();
-        }
-
-        #[tokio::test]
-        #[serial]
-        async fn saves_to_cache() {
-            let dir = assert_fs::TempDir::new().unwrap();
-            let cache = CacheEngine::create(dir.path()).await.unwrap();
-            let mut item = cache.cache_projects_state().await.unwrap();
-
-            item.item
-                .projects
-                .insert("foo".to_owned(), "bar".to_owned());
-
-            run_with_env("", || item.save()).await.unwrap();
-
-            assert_eq!(
-                fs::read_to_string(item.path).unwrap(),
-                r#"{"globs":[],"projects":{"foo":"bar"}}"#
-            );
-
-            dir.close().unwrap();
-        }
-    }
-
-    mod save_hash {
-        use super::*;
-        use serde::Deserialize;
-
-        #[derive(Default, Deserialize, Serialize)]
-        struct TestHasher {
-            field: String,
-        }
-
-        #[tokio::test]
-        #[serial]
-        async fn creates_hash_file() {
-            let dir = assert_fs::TempDir::new().unwrap();
-            let cache = CacheEngine::create(dir.path()).await.unwrap();
-            let hasher = TestHasher::default();
-
-            cache.save_hash("abc123", &hasher).await.unwrap();
-
-            assert!(cache.hashes_dir.join("abc123.json").exists());
-
-            dir.close().unwrap();
-        }
-
-        #[tokio::test]
-        #[serial]
-        async fn doesnt_create_if_cache_off() {
-            let dir = assert_fs::TempDir::new().unwrap();
-            let cache = CacheEngine::create(dir.path()).await.unwrap();
-            let hasher = TestHasher::default();
-
-            run_with_env("off", || cache.save_hash("abc123", &hasher))
-                .await
-                .unwrap();
-
-            assert!(!cache.hashes_dir.join("abc123.json").exists());
-
-            dir.close().unwrap();
-        }
-
-        #[tokio::test]
-        #[serial]
-        async fn doesnt_create_if_cache_readonly() {
-            let dir = assert_fs::TempDir::new().unwrap();
-            let cache = CacheEngine::create(dir.path()).await.unwrap();
-            let hasher = TestHasher::default();
-
-            run_with_env("read", || cache.save_hash("abc123", &hasher))
-                .await
-                .unwrap();
-
-            assert!(!cache.hashes_dir.join("abc123.json").exists());
-
-            dir.close().unwrap();
-        }
     }
 }
